refactor(channels): move discord_rest_client -> channels/discord/client
Slice 6, on owner's no-wait directive. All 5 production senders already reach
it via channels.discord.post_gate.make_client, so the importer set is 11 test
loaders + 2 package-internal imports, plus two load-bearing path pins:
notify.py's repo-root finder (existence-of-file discriminator) and the census
allowlist.
