+++ tests/bot2bot-post-length-guard.test.py
@@ -105,7 +105,7 @@ def main() -> int:
 
     # The PRODUCTION DiscordRestClient stays in the loop; only its transport
     # is scripted, so "did not send" is observed at the real chokepoint.
-    from discord_rest_client import DiscordRestClient
+    from channels.discord.client import DiscordRestClient
     post_mod._client = lambda token: DiscordRestClient(
         token, transport=_recording_transport)
     try:
